+++ src/SCCM.Core/ControlManagerBase.cs
@@ -9,8 +9,8 @@ public interface IControlManager
     event Action<string> DebugOutput;
 
     string GameType { get; }
-    string ReadLocation { get; set; }
-    string SaveLocation { get; set; }
+    string GameConfigLocation { get; set; }
+    string AppSaveLocation { get; set; }
 
     Task Import(ImportMode mode);
     Task ExportPreview();
@@ -31,15 +31,40 @@ public abstract class ControlManagerBase : IControlManager
     protected void WriteLineDebug(string s) => this.DebugOutput(s);
 
     public abstract string GameType { get; }
-    public string ReadLocation { get; set; } = string.Empty;
-    public string SaveLocation { get; set; } = string.Empty;
+    public string GameConfigLocation { get; set; } = string.Empty;
+    public string AppSaveLocation { get; set; } = string.Empty;
+
+    protected abstract string GameConfigPath { get; }
+    protected abstract string MappingDataSavePath { get; }
+    protected IPlatform Platform { get; init; }
+
+    protected ControlManagerBase(IPlatform platform)
+    {
+        this.Platform = platform;
+    }
 
     protected abstract IMappingImporter CreateImporter();
     protected abstract IMappingImportMerger CreateMerger();
     protected abstract IMappingExporter CreateExporter();
 
-    protected abstract Task<MappingData?> LoadMappingData();
-    protected abstract Task SaveMappingData(MappingData data);
+    protected async Task<MappingData?> LoadMappingData()
+    {
+        var serializer = new DataSerializer(this.MappingDataSavePath);
+        return await serializer.Read();
+    }
+
+    protected async Task SaveMappingData(MappingData data)
+    {
+        if (data == null)
+        {
+            throw new ArgumentNullException(nameof(data));
+        }
+
+        System.IO.Directory.CreateDirectory(this.AppSaveLocation);
+        var serializer = new DataSerializer(this.MappingDataSavePath);
+        await serializer.Write(data);
+        WriteLineStandard($"Mappings backed up to [{this.MappingDataSavePath}].");
+    }
 
     public async Task Import(ImportMode mode)
     {
@@ -116,91 +141,17 @@ public abstract class ControlManagerBase : IControlManager
         WriteLineStandard($"{this.GameType} config restored from [{backup}].");
     }
 
-    public abstract void Open();
-
-    public abstract void OpenGameConfig();
-}
-
-public class ControlManager : ControlManagerBase
-{
-    public string StarCitizenActionmapsXmlPath { get => System.IO.Path.Combine(this.ReadLocation, Constants.SC_ACTIONMAPS_XML_NAME); }
-    public string SccmMappingsJsonPath { get => System.IO.Path.Combine(this.SaveLocation, Constants.SCCM_SCMAPPINGS_JSON_NAME); }
-
-    public override string GameType => "Star Citizen";
-
-    private readonly IPlatform _platform;
-    private readonly ISCFolders _folders;
-
-    public ControlManager(IPlatform platform, ISCFolders folders)
-    {
-        this._platform = platform;
-        this._folders = folders;
-        Initialize();
-    }
-
-    private void Initialize()
-    {
-        this.ReadLocation = this._folders.ActionMapsDir;
-        this.SaveLocation = this._folders.SccmDir;
-    }
-
-    protected override MappingImporter CreateImporter()
-    {
-        var importer = new MappingImporter(this._platform, this.StarCitizenActionmapsXmlPath);
-        importer.StandardOutput += WriteLineStandard;
-        importer.WarningOutput += WriteLineWarning;
-        importer.DebugOutput += WriteLineDebug;
-        return importer;
-    }
-
-    protected override MappingImportMerger CreateMerger()
-    {
-        var merger = new MappingImportMerger();
-        merger.StandardOutput += WriteLineStandard;
-        merger.WarningOutput += WriteLineWarning;
-        merger.DebugOutput += WriteLineDebug;
-        return merger;
-    }
-
-    protected override MappingExporter CreateExporter()
-    {
-        var exporter = new MappingExporter(this._platform, this._folders, StarCitizenActionmapsXmlPath);
-        exporter.StandardOutput += WriteLineStandard;
-        exporter.WarningOutput += WriteLineWarning;
-        exporter.DebugOutput += WriteLineDebug;
-        return exporter;
-    }
-
-    protected override async Task<MappingData?> LoadMappingData()
-    {
-        var serializer = new DataSerializer(this.SccmMappingsJsonPath);
-        return await serializer.Read();
-    }
-
-    protected override async Task SaveMappingData(MappingData data)
-    {
-        if (data == null)
-        {
-            throw new ArgumentNullException(nameof(data));
-        }
-
-        System.IO.Directory.CreateDirectory(this.SaveLocation);
-        var serializer = new DataSerializer(this.SccmMappingsJsonPath);
-        await serializer.Write(data);
-        WriteLineStandard($"Mappings backed up to [{this.SccmMappingsJsonPath}].");
-    }
-
-    public override void Open()
+    public void Open()
     {
         // TODO write simple test
-        this._platform.Open(this.SccmMappingsJsonPath);
-        WriteLineStandard($"Opening [{this.SccmMappingsJsonPath}] in the default editor, change the Preserve property to choose which settings are overwritten.");
+        this.Platform.Open(this.MappingDataSavePath);
+        WriteLineStandard($"Opening [{this.MappingDataSavePath}] in the default editor, change the Preserve property to choose which settings are overwritten.");
     }
-    
-    public override void OpenGameConfig()
+
+    public void OpenGameConfig()
     {
         // TODO write simple test
-        this._platform.Open(this.StarCitizenActionmapsXmlPath);
-        WriteLineStandard($"Opening [{this.StarCitizenActionmapsXmlPath}] in the default editor.");
+        this.Platform.Open(this.GameConfigPath);
+        WriteLineStandard($"Opening [{this.GameConfigPath}] in the default editor.");
     }
 }
