+++ src/SCCM.Core/IMappingExporter.cs
@@ -1,0 +1,15 @@
+namespace SCCM.Core;
+
+public interface IMappingExporter
+{
+    event Action<string> StandardOutput;
+    event Action<string> WarningOutput;
+    event Action<string> DebugOutput;
+
+    string GameConfigPath { get; }
+
+    string Backup();
+    string RestoreLatest();
+    Task Preview(MappingData source);
+    Task Update(MappingData source);
+}
