+++ src/SCCM.Core/IMappingImporter.cs
@@ -1,0 +1,10 @@
+﻿namespace SCCM.Core;
+
+public interface IMappingImporter
+{
+    event Action<string> StandardOutput;
+    event Action<string> WarningOutput;
+    event Action<string> DebugOutput;
+
+    Task<MappingData> Read();
+}
