+++ src/SCCM.Core/SC/Constants.cs
@@ -1,4 +1,4 @@
-namespace SCCM.Core;
+namespace SCCM.Core.SC;
 
 public static class Constants
 {
