+++ src/SCCM.Core/SC/MappingImportMerger.cs
@@ -1,17 +1,4 @@
-namespace SCCM.Core;
-
-public interface IMappingImportMerger
-{
-    event Action<string> StandardOutput;
-    event Action<string> WarningOutput;
-    event Action<string> DebugOutput;
-
-    MappingMergeResult Result { get; }
-
-    bool Preview(MappingData current, MappingData updated);
-    
-    MappingData Merge(MappingData current, MappingData updated);
-}
+namespace SCCM.Core.SC;
 
 public class MappingImportMerger : IMappingImportMerger
 {
