+++ src/SCCM.Tests/MappingExporter_Update_Tests.cs
@@ -1,10 +1,12 @@
 using NUnit.Framework;
 using SCCM.Core;
+using SCCM.Core.SC;
 using SCCM.Tests.Mocks;
 using System.Text;
 using System.Xml;
 using System.Xml.Linq;
 using System.Xml.XPath;
+using static SCCM.Core.SC.Extensions;
 using static SCCM.Tests.Extensions;
 
 namespace SCCM.Tests;
